requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.15.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
Brotli==1.1.0
Cerberus==1.3.5
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap4==2024.1
cryptography==43.0.0
cssselect2==0.7.0
daphne==4.1.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.15
django-autoslug==1.9.9
django-autoslug-field==0.2.3
django-bootstrap-icons==0.9.0
django-bootstrap-v5==1.0.11
django-bootstrap4==24.3
django-bootstrap5==24.3
django-chartjs==2.3.0
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-crispy-forms==2.2
django-easy-audit==1.3.6
django-environ==0.11.2
django-feather==0.3.0
django-filter==24.2
django-formset==1.4.5
django-import-export==4.0.9
django-jquery==3.1.0
django-js-asset==2.2.0
django-model-utils==4.5.1
django-money==3.5.2
django-notifications-hq==1.8.3
django-pdfkit==0.3.1
django-permissions-policy==4.19.0
django-phonenumber-field==7.3.0
django-selectize==2.0
django-weasyprint==2.3.0
django_csp==3.8
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
fontawesome-free==5.15.4
fontawesomefree==6.5.1
fonttools==4.53.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonfield==3.1.0
MarkupPy==1.14
MarkupSafe==2.1.5
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.0
phonenumbers==8.13.39
phonenumberslite==8.13.40
pillow==10.4.0
protobuf==5.26.0
pycparser==2.22
pydyf==0.10.0
pysodium==0.7.17
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.4
reportlab==4.2.2
requests==2.31.0
service-identity==24.1.0
shortuuid==1.0.13
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
swapper==1.4.0
tablib==3.5.0
tinycss2==1.3.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.1
user-agents==2.2.0
weasyprint==62.2
webencodings==0.5.1
zope.interface==6.4.post2
zopfli==0.2.3
django-tinymce==4.1.0